Split out ProjectAvailabilityTests
